fix: refactor module structure; fix loss argument not working in fnn trainer agent
