+++ problem-solver/py/modules/fnnProcessingModule/FnnInterpreterModule.py
@@ -1,0 +1,7 @@
+from sc_kpm import ScModule
+from .FnnInterpreterAgent import FnnInterpreterAgent
+
+
+class FnnInterpreterModule(ScModule):
+    def __init__(self):
+        super().__init__(FnnInterpreterAgent())
