--- problem-solver/py/modules/messageProcessingModule/FnnTrainerProcessingModule.py
@@ -1,7 +1,0 @@
-from sc_kpm import ScModule
-from .FnnTrainerAgent import AgentTraining
-
-
-class FnnTrainerProcessingModule(ScModule):
-    def __init__(self):
-        super().__init__(AgentTraining())
