+++ problem-solver/py/server.py
@@ -1,7 +1,7 @@
 import argparse
 from sc_kpm import ScServer
-from modules.messageProcessingModule.FnnAgentProcessingModule import FnnAgentProcessingModule
-from modules.messageProcessingModule.FnnTrainerProcessingModule import FnnTrainerProcessingModule
+from modules.fnnProcessingModule.FnnInterpreterModule import FnnInterpreterModule
+from modules.fnnProcessingModule.FnnTrainingModule import FnnTrainingModule
 from pathlib import Path
 
 SC_SERVER_PROTOCOL = "protocol"
@@ -19,8 +19,8 @@ def main(args: dict):
 
     with server.connect():
         modules = [
-            FnnAgentProcessingModule(),
-            FnnTrainerProcessingModule()
+            FnnInterpreterModule(),
+            FnnTrainingModule()
         ]
         server.add_modules(*modules)
         with server.register_modules():
